* re-factored and consolidated code generator * generated 0.4.6 PersonIdentification schema and java sdk components

+++ soi-track/src/main/resources/tool-bin/soi-java-sdk-schema-binder/bin/unzip-source.sh
@@ -1,9 +1,16 @@
-TARGET_DIR="source"
-#ZIP_FILE="TsoaTrack-01.zip"
-ZIP_FILE=${1}
+#!/bin/bash
 
-#rm -rvf source/iep-schema/*
-rm -rvf ${TARGET_DIR}/*
-#unzip -o TsoaTrack-01.zip -d source/iep-schema
-unzip -o ${ZIP_FILE} -d ${TARGET_DIR}
-rm -vf ${TARGET_DIR}/iep-schema/TsoaTrack-dereferenced.json
+if [ ${1} > 0 ] ; then
+  set -x
+  TARGET_DIR="source"
+  #ZIP_FILE="TsoaTrack-01.zip"
+  ZIP_FILE=${1}
+
+  #rm -rvf source/iep-schema/*
+  rm -rvf ${TARGET_DIR}/*
+  #unzip -o TsoaTrack-01.zip -d source/iep-schema
+  unzip -o ${ZIP_FILE} -d ${TARGET_DIR}
+  rm -vf ${TARGET_DIR}/iep-schema/TsoaTrack-dereferenced.json
+else
+  echo "  usage: ${0} ZIP_FILE"
+fi
